move user function in user.ts and save current user in users array in local storage
