+++ src/pages/profile_seeker/Preferences.vue
@@ -1,7 +1,7 @@
 <template>
     <p>Wähle aus, welche Tiere du suchen möchtest.</p>
     <div class="container">
-        <div class="row g-2">
+        <div class="row g-2" v-if="currentUser">
             <div v-for="animal of currentUser.preferences" class="col-6">
                 <div class="card" role="button" :class="{ 'bg-primary': animal.active }"
                     @click="animal.active = !animal.active">
@@ -15,17 +15,6 @@
 
 </template>
 <script setup lang="ts">
-import { ref, watch } from 'vue';
-
-const currentUser = ref(getCurrentUser())
-watch(currentUser, setCurrentUser, { deep: true });
-
-function setCurrentUser() {
-    localStorage.setItem('current_user', JSON.stringify(currentUser.value));
-}
-function getCurrentUser() {
-    return JSON.parse(localStorage.getItem('current_user') || "false");
-}
-
+import { currentUser } from "../../user";
 </script>
 <style></style>
